requirements.txt
spotipy
python-dotenv
devtools
